yank: detect tmux via $TERM=screen* in nested SSH
When you start an SSH session from inside tmux, the $TMUX environment
variable does not carry forward into that newly started SSH session.
As a result, when you run the yank script inside that SSH session, it
doesn't know that it's actually attached to a tmux client (since $TMUX
is not set) and it therefore doesn't wrap the OSC 52 sequence for tmux.
Consequently, tmux discards the OSC 52 sequence and it all falls apart.

To solve this, we fall back to the $TERM environment variable (which
begins with the word "screen" for all terminals started under tmux).